mcp_servers_workflows/requirements.txt
# MCP Server Requirements
fastmcp>=2.10.0
httpx>=0.24.0
pydantic>=2.0.0
starlette>=0.37.0
temporalio>=1.8.0
python-dotenv>=1.0.0